% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{Visualize earthquakes data on leaflet Map.}
\usage{
eq_map(data, annotation)
}
\arguments{
\item{data}{Filtered dataframe of NOAA earthquakes dataset}

\item{annotation}{Column name from dataset used for the annotation in the pop-up}
}
\value{
leaflet MAP shown with a circle markers, and the radius of the circle marker is
         proportional to the earthquake's magnitude.
}
\description{
This akes an argument data containing the filtered data frame with earthquakes to visualize.
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point with in pop up
window containing annotation data stored in a column of the data frame.
}
\examples{
\dontrun{
 eq_map(data, annotation = "DATE")

 data \%>\%
       dplyr::filter(COUNTRY \%in\% COUNTRIES & lubridate::year(DATE) >= 2000) \%>\%
       eq_map(annotation = "DATE")
}

}
