% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{theme_timeline}
\alias{theme_timeline}
\title{A theme for styling \code{geom_timeline} plots}
\usage{
theme_timeline()
}
\description{
A theme for styling \code{\link{geom_timeline}} plots.
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
filter(COUNTRY \%in\% c("USA", "CHINA"), YEAR > 2000) \%>\%
   ggplot(aes(x = DATE,
              y = COUNTRY,
              color = as.numeric(TOTAL_DEATHS),
              size = as.numeric(EQ_PRIMARY)
   )) +
   geom_timeline() +
   theme_timeline()
}

}
