% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Popup annotation for \code{eq_map} with Location, Magnitude, and Total number of Deaths}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Filtered dataframe of NOAA earthquakes dataset}
}
\value{
Popup text for Location, Magnitude, and Total number of Deaths (HTML formatted)
}
\description{
Creates an annotation for each earthquake that will show the location,
magnitude (EQ_PRIMARY), and total number of deaths (TOTAL_DEATHS), with respective
labels "Location", "Total deaths", and "Magnitude". If an
earthquake is missing values for any of these, both the label and the value
will be skipped for that element of the tag.
}
\examples{
\dontrun{
 eq_create_label(data)

 data \%>\%
       dplyr::filter(COUNTRY \%in\% COUNTRIES & YEAR >= 2000) \%>\%
       dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
       eq_map(annotation = "popup_text")
}

}
