% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{Visualize earthquakes data on leaflet Map.}
\usage{
eq_map(data, annotation)
}
\arguments{
\item{data}{Filtered dataframe of NOAA earthquakes dataset}

\item{annotation}{Column name from dataset used as annotation in the popup}
}
\value{
leaflet MAP shown with a circle markers, and the radius of the circle marker is
         proportional to the earthquake's magnitude.
}
\description{
Takes a clean NOAA data frame with earthquakes to visualize. Maps the
epicenters (LATITUDE/LONGITUDE) and annotates each earthquake with a popup.
}
\examples{
\dontrun{
 eq_map(data, annotation = "DATE")

 data \%>\%
       dplyr::filter(COUNTRY \%in\% COUNTRIES & YEAR >= 2000) \%>\%
       eq_map(annotation = "DATE")
}

}
